handle to cpplint

+++ src/chapter01/chapter01.h
@@ -10,13 +10,14 @@
 #include <iostream>
 #include <list>
 #include <random>
-#include <regex>
+#include <regex>  // NOLINT
+#include <string>
 #include <utility>
 #include <vector>
 
 
 namespace chapter01 {
-unsigned long long q1(unsigned long long);
+unsigned int q1(int);
 int q2(unsigned int, unsigned int);
 int q3(unsigned int, unsigned int);
 int q4(int);
@@ -24,12 +25,12 @@ std::vector<std::pair<int, int>> q5(int);
 std::vector<std::pair<int, int>> q6(int);
 std::vector<std::pair<int, int>> q7();
 std::vector<int> q8();
-std::vector<std::pair<int,int>> q9(int);
-std::vector<std::pair<std::bitset<5>,std::bitset<5>>> q10();
-std::string q11(unsigned long long);
+std::vector<std::pair<int, int>> q9(int);
+std::vector<std::pair<std::bitset<5>, std::bitset<5>>> q10();
+std::string q11(unsigned int);
 int q12();
 double q13();
 bool q14(const std::string&);
-}
+}  // namespace chapter01
 
 #endif  // SRC_CHAPTER01_CHAPTER01_H_
